+++ CHANGELOG.md
@@ -7,6 +7,51 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.6](https://github.com/instantOS/instantCLI/compare/v0.2.5...v0.2.6) - 2025-10-17
+
+### Fixed
+
+- fix CI
+
+## [0.2.5](https://github.com/instantOS/instantCLI/compare/v0.2.4...v0.2.5) - 2025-10-17
+
+### Fixed
+
+- fix typo
+
+### Other
+
+- yamlfmt
+
+## [0.2.4](https://github.com/instantOS/instantCLI/compare/v0.2.3...v0.2.4) - 2025-10-17
+
+### Fixed
+
+- fix appimage build
+
+## [0.2.3](https://github.com/instantOS/instantCLI/compare/v0.2.2...v0.2.3) - 2025-10-17
+
+### Fixed
+
+- fix appimage
+
+### Other
+
+- add appimage to CI
+- change appimage icon
+- simplify appimage
+- init appimage build
+- desus christ
+
+## [0.2.2](https://github.com/instantOS/instantCLI/compare/v0.2.1...v0.2.2) - 2025-10-15
+
+### Other
+
+- init binstall compatibility
+- clippy
+- format
+- add comment tests
+
 ## [0.2.1](https://github.com/instantOS/instantCLI/compare/v0.2.0...v0.2.1) - 2025-10-14
 
 ### Other